requirements.txt
Flask==2.0.2
Flask_Login==0.5.0
Flask_SQLAlchemy==2.5.1
fpdf==1.7.2
fpdf2==2.4.6
python-dotenv==0.20.0
qrcode==7.3.1
SQLAlchemy_serializer==1.4.1
Werkzeug==2.0.2
